+++ packages/core/src/atoms/index.ts
@@ -4,3 +4,4 @@
 
 export * from './createAtomStore';
 export * from './elementAtom';
+export * from './useElement';
